+++ api/src/main/java/org/allaymc/api/block/interfaces/BlockStandingBannerBehavior.java
@@ -1,6 +1,8 @@
 package org.allaymc.api.block.interfaces;
 
 import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.component.BlockBlockEntityHolderComponent;
+import org.allaymc.api.blockentity.interfaces.BlockEntityBanner;
 
-public interface BlockStandingBannerBehavior extends BlockBehavior {
+public interface BlockStandingBannerBehavior extends BlockBehavior, BlockBlockEntityHolderComponent<BlockEntityBanner> {
 }