+++ api/src/main/java/org/allaymc/api/blockentity/interfaces/BlockEntityBanner.java
@@ -1,0 +1,12 @@
+package org.allaymc.api.blockentity.interfaces;
+
+import org.allaymc.api.blockentity.BlockEntity;
+import org.allaymc.api.blockentity.component.BlockEntityBannerBaseComponent;
+
+/**
+ * Represents a banner block entity.
+ *
+ * @author daoge_cmd
+ */
+public interface BlockEntityBanner extends BlockEntity, BlockEntityBannerBaseComponent {
+}