+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemBannerStack.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.component.ItemBannerBaseComponent;
 
-public interface ItemBannerStack extends ItemStack {
+public interface ItemBannerStack extends ItemStack, ItemBannerBaseComponent {
 }